Проблема при добавлении метода mapRowToReview

+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class Review {
-    private Long reviewId;
+    private int reviewId;
     @NotBlank(message = "Содержание отзыва не может быть пустым")
     private String content;
     private boolean isPositive;
